+++ examples/resources/iosxr_banner/resource.tf
@@ -1,4 +1,4 @@
 resource "iosxr_banner" "example" {
-  banner_type = "prompt-timeout"
-  line        = ", banner-text ,"
+  banner_type = "login"
+  line        = " Hello user  !"
 }